src/app/components/07-forms/user-form/user-form.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { UserFormComponent } from './user-form.component';
import { MatInputModule } from '@angular/material/input';
import { FormsModule, ReactiveFormsModule } from '@angular/forms';
import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
import { MatButtonModule } from '@angular/material/button';

describe('UserFormComponent', () => {
  let component: UserFormComponent;
  let fixture: ComponentFixture<UserFormComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [UserFormComponent],
      imports: [MatInputModule, ReactiveFormsModule, FormsModule, BrowserAnimationsModule, MatButtonModule]
    })
      .compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(UserFormComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should save', () => {
    spyOn(component.userCreated, "emit");
    spyOn(component.userForm, "reset");
    component.userForm.get('name')?.setValue("test");
    component.save();
    expect(component.userCreated.emit).toHaveBeenCalled();
    expect(component.userForm.reset).toHaveBeenCalled();
  });
});
